Move text generation to three-d
